修改content-length的存储类型为u64，保证大文件可以正常下载 （cherry picked commit from <gitee.com//openharmony/commonlibrary_rust_ylong_http/commit/3e661801976f420ec9d82a376523c0ef600884e8>

+++ ylong_http_client/src/async_impl/http_body.rs
@@ -275,7 +275,7 @@ struct Text {
 
 impl Text {
     pub(crate) fn new(
-        len: usize,
+        len: u64,
         pre: &[u8],
         io: BoxStreamData,
         interceptors: Arc<Interceptors>,
